# perf(trace): avoid to allocate option struct in span tracking